Printid välja integeri

+++ newHope/Display.java
@@ -69,7 +69,7 @@ public class Display{
     }
 
     public static String generateLine() {
-        return "Vanus:"+ User.age+" Sõbrad:"+ User.friends+" Teadmised:"+ User.knowledge+" Raha:"+User.money+
+        return "Vanus:"+ (int)User.age+" Sõbrad:"+ (int)User.friends+" Teadmised:"+ (int)User.knowledge+" Raha:"+ (int)User.money+
         		" Mõistus:"+Math.round(User.sanity*100)+" Kõht täis:"+Math.round(User.hunger*100)+
                 " Tervis:"+Math.round(User.health*100);
     }
